bot/requirements.txt
discord.py
slack_sdk
python-dotenv
psycopg2-binary
matplotlib
japanize-matplotlib
pytz
